test(e2e): rename the live-backend suite from integration to backend
