+++ e2e/backend/global-setup.ts
@@ -5,7 +5,7 @@ import { ping } from '../helpers/temp-email.js'
 export async function setup() {
   if (!process.env.ZD_PROJECT_ID || !process.env.ZD_OTP_PROJECT_ID) {
     throw new Error(
-      'ZD_PROJECT_ID and ZD_OTP_PROJECT_ID are required for staging integration tests',
+      'ZD_PROJECT_ID and ZD_OTP_PROJECT_ID are required for the staging backend e2e tests',
     )
   }
