+++ package.json
@@ -21,10 +21,10 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
-    "test:integration": "vitest run --config e2e/vitest.integration.config.ts",
     "build:e2e": "pnpm --filter '@zerodev/qa-lab-testing^...' build",
-    "test:e2e": "pnpm build:e2e && playwright test --config e2e/playwright.config.ts",
-    "test:e2e:headed": "pnpm build:e2e && playwright test --config e2e/playwright.config.ts --headed",
+    "test:e2e:browser": "pnpm build:e2e && playwright test --config e2e/playwright.config.ts",
+    "test:e2e:browser:headed": "pnpm build:e2e && playwright test --config e2e/playwright.config.ts --headed",
+    "test:e2e:backend": "vitest run --config e2e/vitest.backend.config.ts",
     "check:publishable": "node scripts/check-publishable.mjs",
     "prepare": "husky",
     "changeset": "changeset",
